backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
langgraph
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
passlib[bcrypt]>=1.7
bcrypt==3.2.2